refactor base artifact to a separate file
